Improve ReusedModifierInstance
This rule now uses type resolution to be smarter and is now able to
correctly detekt when a composable call does not have a modifier
parameter at all, and will not report usage of `modifier` in composable
children of this call as an error.

Fixes #5

Also added a function which appends a "fake" compose
functions/classes/imports to mimic ones provided by Compose framework.

This is required, because at the moment Detekt doesn't support adding
custom plugins when it builds BindingContext and while Compose libraries
can be added on classpath through testImplementation in build.gradle,
but snippets won't compile without the Compose plugin.