Add new rule: `ConditionCouldBeLifted`
It will detect cases when if-condition inside a composable layout call
could be "lifted up" and the whole call could be moved into that
conditional expression, for example:

```
Column {
  if (x == 3) {
    Text("1")
    Text("2")
  }
}
```

could be turned into

```
if (x == 3) {
  Column {
    Text("1")
    Text("2")
  }
}
```

At the moment it tries to be extra careful to avoid reporting any
potentially side-effecting code (for example if `Column` in the example
above would have some `Modifier` affecting a parent layout, this
conditional-lifting change wouldn't be correct), and by being "extra"
careful it can miss some potential cases for optimisation.

This may be improved in future.